Notify FileTransferStatusListeners on each transferred chunk
For debug reasons (and for simplified GUI updates, i. e. push vs
poll) it makes sense to provide parties interested in file transfer
status not only the overall status (IN_PROGRESS, FAILED, CLOSED) but
also the amount of transferred bytes.

By intention the progress percentage is not contained in the event, as
it is only a function of bytes transferred and length, hence, would be
a redundant information.