Storing session produced by createSession
A chaat session produced by createSession was not stored in the session
cache. As a result, as soon as the chat partner sends his first message
(possibly also an eagerly sent standalone chat state notification), you
end up having two sessions with the same partner -- one in the cache,
one not in the cache.

This is weird, as it gives outgoing created sessions a special state
which is not documented anywhere. All session should be the same. This
commit fixes this, as it registers outgoing sessions just the same way
as incoming sessions.

As a side effect it becomes apparent that there is an ongoing design
fault in session management: What is a session at all and how to start
and end it? This question is to be solved in a separate commit, and a
proposal could be to simply not send a <thread> id with outgoing session
and just wait until the partner answers -- which will assign a <thread>
id THEN.

Possibly this was wanted originally, then the current commit is
counterproductive but should get replaced by one that actually fixes the
problem that outgoing sessions currently do have a thread id!