Don't log exceptions too eagerly.
Exceptions which occur during writing to the socket cause the connection to get closed.
Exceptions which in turn occur during closing the connection should be added as suppressed exceptions.

Also make sure that non-final variables become null, so that they are not shutdown/closed twice (which will likely cause a new exception).

Fixes Issue #19