Eagerly releases unused port to prevent ports-leaks due delayed GC
As long as a socket was bound once, it reserves a TCP port of the local
host. This port is freed either explicitly on close() or implicitly at
GC. In some cases it can happen that GC is delayed for long time (e. g.
when the host has plenty of RAM and / or has only very few cores). As a
consequence, unused sockets eat up ports, which are a limited resource.
There is a risk to run out of ports eventually, particularly on
long-running services powered by shared resources (e. g. virtualization,
containers, etc.).

This patch reduces this risk by eagerly releasing TCP ports of sockets
which became "unusable" in the sense of Babbler (open but disconnected)
instead of simply entitle them for GC by not holding references anymore.

The patch possibly does not completely remedy the risk as for that
purpose the existing source code has to be checked for more situations
in which a previously used socket is getting unreferenced unclosed.