Correctly deals with missed reconnection event due to suspended host.
When the host got suspended for a duration being longer than the planned
delay, certainly the attempt has to be executed AT ONCE when the host is
getting resumed again. Otherwise the paradox situation would happen that
the actual delay equals the planned delay PLUS the suspend time, which
would be simply weird as it semantically implies an additional penalty
for already having waited longer than actually requested...!

The cause of the evil is that the Oracle JVM does not wait for an
absolute given target time to be reached, but instead it seems to
internally count the number of ticks. While the PC is suspended, no
ticks are sent to the JVM, hence it does not notice how fast or slow
absolute time passes by meanwhile. Hence, when it wakes up, it has
"seen" the same number of ticks as before it fell asleep.