Connection's send() should return CompletableFuture instead of Future.
This commit reworks the semantics of the Future returned by the send method.

Previously the Future simply was always (successfully) done, when the send task has been processed, even if there was an exception.
Now the Future fails, if the sending fails.

This allows us to hook callbacks, i.e. get notified if a send failed.

Now, the returned Future isDone(), when the stanza has been successfully sent.