Resolve rare deadlock when using Stream Management
This deadlock only occurred when using stream management and only when a message is sent after receiving a request (<r/>) from the server but before the answer (<a/>) is sent.
The time frame for this is very small, but in rare cases it could still occur (e.g. it took several days to happen).

Also generally make sure, that the send method is not called from within synchronized blocks.