Update the session status before closing the connection
Otherwise closing may throw an exception, and updates the status with that exception with the consequence that the original exception is never seen.

Concrete case:
Openfire < 4.2 closes the connection without </stream:stream>. When preceeded by a stream error, it won't be notified, because closing the connection first results in a SocketException, which actually closes the session.