Introduce a new way to establish a connection
This commit adds the concept of a "transport connector", which allows for different implementations for a binding to a transport protocol.

E.g. the TCP binding could be implemented using either java.net.Sockets or Netty Channels.

This approach feels more advantageous and sustainable than introducing more and more derivations of TcpConnectionConfiguration.